+++ shell.core/src/main/java/org/metaborg/spoofax/shell/core/CoreModule.java
@@ -16,7 +16,7 @@ public class CoreModule extends AbstractModule {
     @Override
     protected void configure() {
         bind(IReplCommand.class).annotatedWith(Names.named("EvalCommand"))
-                .to(SpoofaxEvaluationCommand.class).in(Singleton.class);
+            .to(SpoofaxEvaluationCommand.class).in(Singleton.class);
         bind(ICommandInvoker.class).to(SpoofaxCommandInvoker.class);
     }
 }
